Move analysis files to enable further uses of the repository
