+++ src/generate_analysis_files/reports/workplace/index.js
